+++ lib/orian/rs.ex
@@ -9,4 +9,9 @@ defmodule Orian.Rs do
 
   def blake3(_bin), do: :erlang.nif_error(:nif_not_loaded)
   def xxh3(_bin), do: :erlang.nif_error(:nif_not_loaded)
+  def bulk(_jobs, _conc), do: :erlang.nif_error(:nif_not_loaded)
+  def put_file(_url, _path, _headers), do: :erlang.nif_error(:nif_not_loaded)
+  def get_file(_url, _path, _headers), do: :erlang.nif_error(:nif_not_loaded)
+  def pipe(_su, _sh, _du, _dh), do: :erlang.nif_error(:nif_not_loaded)
+  def engine_loaded, do: false
 end